pyproject.toml
[tool.poetry]
name = "grouper"
version = "0.1.0"
description = ""
authors = ["lamar1524 <pawel.wybrzeze@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.6"
djangorestframework = "^3.11.0"
Pillow = "^7.1.2"
django-cors-headers = "^3.3.0"
daphne = "^2.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = "^5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
